+++ src/main/java/com/austinmreppert/graphio/GraphIO.java
@@ -3,7 +3,9 @@ package com.austinmreppert.graphio;
 import com.austinmreppert.graphio.capabilities.Capabilities;
 import com.austinmreppert.graphio.network.PacketHandler;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
@@ -20,6 +22,8 @@ public class GraphIO {
 
   public GraphIO() {
     FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
+
+    ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, Config.SERVER_SPEC);
 
     MinecraftForge.EVENT_BUS.register(this);
   }